move encodings into js, port tests to jest
