--- encodings/fastpfor/js/.cspell.json
@@ -1,34 +1,0 @@
-{
-  "version": "0.1",
-  "$schema": "https://raw.githubusercontent.com/streetsidesoftware/cspell/master/cspell.schema.json",
-  "language": "en",
-  "words": [
-    "bitjson",
-    "bitauth",
-    "cimg",
-    "circleci",
-    "codecov",
-    "commitlint",
-    "dependabot",
-    "editorconfig",
-    "esnext",
-    "execa",
-    "exponentiate",
-    "globby",
-    "libauth",
-    "mkdir",
-    "prettierignore",
-    "sandboxed",
-    "transpiled",
-    "typedoc",
-    "untracked"
-  ],
-  "flagWords": [],
-  "ignorePaths": [
-    "package.json",
-    "package-lock.json",
-    "yarn.lock",
-    "tsconfig.json",
-    "node_modules/**"
-  ]
-}
