--- encodings/fastpfor/js/.gitignore
@@ -1,10 +1,0 @@
-diff
-.idea/*
-.nyc_output
-build
-node_modules
-test
-src/**.js
-coverage
-*.log
-yarn.lock
