--- encodings/fastpfor/js/tsconfig.module.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "target": "esnext",
-    "outDir": "build/module",
-    "module": "esnext"
-  },
-  "exclude": [
-    "node_modules/**"
-  ]
-}
