--- encodings/fsst/js/.gitignore
@@ -1,6 +1,0 @@
-node_modules/
-dist/
-coverage/
-.vscode/
-.idea/
-*.log
