--- encodings/fsst/js/.prettierrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    semi: true,
-    singleQuote: false,
-    printWidth: 120,
-    tabWidth: 4,
-};
