--- encodings/fsst/js/README.md
@@ -1,19 +1,0 @@
-# FSST Decoder
-This is a TypeScript project that implements the FSST decoding algorithm.
-
-## Getting Started
-### Prerequisites
-- Node.js and npm
-- TypeScript
-
-### Installation
-1. `cd encodings/fsst/js`
-2. `npm install` to install the dependencies
-3. `npm run build` to compile the project
-4. `npm run start` to run the project
-
-### Usage
-Check the `src/index.ts` file for an example of how to use the FSST decoder.
-
-### Testing
-- `npm test` to run the tests
