--- encodings/fsst/js/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-    transform: {
-        "^.+\\.ts": "ts-jest",
-    },
-    testRegex: "(/test/.*|(\\.|/)(test|spec))\\.(js|ts)$",
-    moduleFileExtensions: ["ts", "js", "json", "node"],
-};
