+++ js/package.json
@@ -26,8 +26,12 @@
   },
   "dependencies": {
     "@mapbox/vector-tile": "^1.3.1",
+    "@types/bytebuffer": "^5.0.49",
+    "@types/varint": "^6.0.3",
+    "bytebuffer": "^5.0.1",
     "ieee754": "^1.2.1",
     "pbf": "^3.2.1",
-    "sqlite3": "^5.1.6"
+    "sqlite3": "^5.1.6",
+    "varint": "^6.0.0"
   }
 }
