+++ js/src/encodings/fastpfor/util.ts
